Promise.all not returned

+++ electron-hack/browser/init.js
@@ -108,15 +108,15 @@ let packageJson = null
 const originalFs = require('original-fs')
 let updateAsar = path.join(process.resourcesPath, 'app-updated.asar')
 let appAsar = path.join(process.resourcesPath, 'app.asar')
-if(originalFs.existsSync(appAsar) && originalFs.existsSync(updateAsar)) {
-    originalFs.unlinkSync(appAsar);
+if(originalFs.existsSync(updateAsar)) {
+    if(originalFs.existsSync(appAsar)) originalFs.unlinkSync(appAsar);
     originalFs.renameSync(updateAsar, appAsar);
     console.log("find app-updated.asar, successfully renamed.");
 }
 updateAsar = path.join(process.resourcesPath, 'lib-updated.asar')
 appAsar = path.join(process.resourcesPath, 'lib.asar')
-if(originalFs.existsSync(appAsar) && originalFs.existsSync(updateAsar)) {
-    originalFs.unlinkSync(appAsar);
+if(originalFs.existsSync(updateAsar)) {
+    if(originalFs.existsSync(appAsar)) originalFs.unlinkSync(appAsar);
     originalFs.renameSync(updateAsar, appAsar);
     console.log("find lib-updated.asar, successfully renamed.");
 }
